perm stuff

+++ mysql_api/src/user/userController.js
@@ -23,8 +23,12 @@ exports.getAllUsers = catchAsync(async (req, res, next) => {
 //user if self || admin
 exports.getUser = catchAsync(async (req, res, next) => {
   const user = await userModel.selectUserById(req.params.id);
+  const userData =
+    !req.userAccess.permissions.includes("admin") && req.userId != req.params.id
+      ? user[0].username
+      : user;
   responseHandler.respond(
-    { head: user[0]?.user_id, data: user },
+    { head: userData[0].username, data: userData },
     { sCode: 200, errCode: 404, errMessage: "User not found" },
     res,
     next
@@ -32,6 +36,13 @@ exports.getUser = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteUser = catchAsync(async (req, res, next) => {
+  if (
+    !req.userAccess.permissions.includes("admin") &&
+    req.userId != req.params.id
+  )
+    return next(
+      new AppError("You have no perrmision to perform this operation", 403)
+    );
   const user = await userModel.deactivateUser(req.params.id);
   responseHandler.respondEmpty(
     { head: user, data: user },
@@ -52,6 +63,13 @@ exports.getUserChallenges = catchAsync(async (req, res, next) => {
 });
 
 exports.getUserChallenge = catchAsync(async (req, res, next) => {
+  if (
+    !req.userAccess.permissions.includes("admin") &&
+    !req.userAccess.createdChallenges.includes(req.params?.challenge)
+  )
+    return next(
+      new AppError("You have no perrmision to perform this operation", 403)
+    );
   const challenge = await userModel.selectUserChallenge(
     req.params.id,
     req.params?.challenge
@@ -65,19 +83,33 @@ exports.getUserChallenge = catchAsync(async (req, res, next) => {
 });
 
 exports.signUpUserChallenge = catchAsync(async (req, res, next) => {
+  if (
+    !req.userAccess.permissions.includes("admin") &&
+    !req.userAccess.createdChallenges.includes(req.params?.challenge)
+  )
+    return next(
+      new AppError("You have no perrmision to perform this operation", 403)
+    );
   const result = await userModel.insertUserChallenge(
     req.params.id,
     req.params?.challenge
   );
   responseHandler.respond(
     { head: +result, data: result },
-    { sCode: 200, errCode: 404, errMessage: result },
+    { sCode: 200, errCode: 500, errMessage: result },
     res,
     next
   );
 });
 
 exports.signOffUserChallenge = catchAsync(async (req, res, next) => {
+  if (
+    !req.userAccess.permissions.includes("admin") &&
+    !req.userAccess.createdChallenges.includes(req.params?.challenge)
+  )
+    return next(
+      new AppError("You have no perrmision to perform this operation", 403)
+    );
   const result = await userModel.deleteUserChallenge(
     req.params.id,
     req.params?.challenge
@@ -91,6 +123,13 @@ exports.signOffUserChallenge = catchAsync(async (req, res, next) => {
 });
 
 exports.modifyUserChallenge = catchAsync(async (req, res, next) => {
+  if (
+    !req.userAccess.permissions.includes("admin") &&
+    req.userId != req.params.id
+  )
+    return next(
+      new AppError("You have no perrmision to perform this operation", 403)
+    );
   const result = await userModel.updateUserChallenge(
     req.body,
     req.params.id,
